src/main/java/kr/megaptera/jdbc/assignment/application/GetPostService.java
package kr.megaptera.jdbc.assignment.application;

import kr.megaptera.jdbc.assignment.daos.PostDao;
import kr.megaptera.jdbc.assignment.domain.Post;
import kr.megaptera.jdbc.assignment.dtos.PostDto;
import kr.megaptera.jdbc.assignment.exceptions.PostNotFound;
import kr.megaptera.jdbc.assignment.repositories.PostRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class GetPostService {

    private final PostDao postDao;

    public PostDto getPost(String id) {
        Post post = postDao.find(id);

        if (post != null) {
            return new PostDto(post);
        } else {
            return null;
        }

    }


}
